:sparkles: 新增小说最新章节信息查询接口

+++ src/main/java/com/mumu/novel/dao/entity/BookChapter.java
@@ -32,7 +32,7 @@ public class BookChapter implements Serializable {
     /**
      * 章节号
      */
-    private Short chapterNum;
+    private Integer chapterNum;
 
     /**
      * 章节名
